delete method_descriptor.go

+++ src/jvmgo/jvm/rtda/class/member_descriptors.go
@@ -18,8 +18,8 @@ type MethodDescriptor struct {
 }
 
 // parameterCount()
-func (self *MethodDescriptor) argCount() (int) {
-    return len(self.parameterTypes)
+func (self *MethodDescriptor) argCount() (uint) {
+    return uint(len(self.parameterTypes))
 }
 
 func (self *MethodDescriptor) addParameterType(t *FieldType) {
